adding settings command, again changed the exception information for maximum detail, shuffled some module commands based on most used

+++ src/main/java/com/basketbandit/core/controllers/GenericMessageController.java
@@ -109,14 +109,17 @@ public class GenericMessageController {
                     clazz = Class.forName(commandModule);
                     constructor = clazz.getConstructor(MessageReceivedEvent.class, String[].class);
 
-                    // Apparently some people aren't giving the bot the permissions they should. This check will let them know.
-                    if(!e.getGuild().getMemberById(420682957007880223L).hasPermission(Permission.MESSAGE_MANAGE)) {
-                        Utils.sendMessage(e,"Sorry, cannot perform action due to a lack of permission. Missing permission: 'MESSAGE_MANAGE'");
-                        return;
-                    } else {
-                        // Remove the input message.
-                        e.getMessage().delete().queue();
-                        break;
+                    // Check settings to see if command strings are to be deleted
+                    if(new DatabaseFunctions().getServerSetting("deleteExecuted", e.getGuild().getId())) {
+                        // Apparently some people aren't giving the bot the permissions they should. This check will let them know.
+                        if(!e.getGuild().getMemberById(420682957007880223L).hasPermission(Permission.MESSAGE_MANAGE)) {
+                            Utils.sendMessage(e, "Sorry, cannot perform action due to a lack of permission. Missing permission: 'MESSAGE_MANAGE'");
+                            return;
+                        } else {
+                            // Remove the input message.
+                            e.getMessage().delete().queue();
+                            break;
+                        }
                     }
                 }
             }
